Refactored single message playing functionality to a separate controller
