+++ lib/voicemail.rb
@@ -7,6 +7,7 @@ require "voicemail/voicemail_controller"
 require "voicemail/mailbox_controller"
 require "voicemail/mailbox_main_menu_controller"
 require "voicemail/mailbox_messages_controller"
+require "voicemail/mailbox_play_message_controller"
 require "voicemail/mailbox_set_greeting_controller"
 require "voicemail/mailbox_set_pin_controller"
